requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
pydantic-settings
python-dotenv
